QtPrivate::equalStrings(): enforce matching size where applicable
Commit 8a7b74f6e71d6b8264f365148bf5a04db4038617 added these functions
and already placed the checking for the size in the inline code. Aside
from the operators==, the only other place that calls equalStrings()
directly is QAnyStringView::equal() and that also checks the size for
the non-UTF-8 string views do match. Unfortunately, the size checking
for UTF-8 was only added in 52e0a91fbc187ca09d3ea6279e8bdce982edc586.

This reverts commit e0eb93d9a2b104ac8d7596c419d1f13dc9db9b1b
("QLatin1StringView: delegate operator== to QByteArrayView") because
that would compare the sizes again.

Change-Id: I83dda2d36c904517b3c0fffd17b38e422b5e2c25
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>