QWidgetRepaintManager: Use textureChildSeen to determine deferred update
We want to get rid of the QWindowPrivate::compositing member, as it is
no longer needed for the Android backingstore, which it was first added
for in a4f50269f82695fbd0dd344f87b4b355feff4333.

We can use textureChildSeen instead to check if we're compositing.

Change-Id: If85b21f92c8253bf89543a7e81e03730023f8095
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
(cherry picked from commit d31cc6dc7b1d1762990fdb70702981f73d060903)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>