Windeployqt: fix exclusion option for Internal modules
The "Internal" suffix was added to module names, but only in the help
text and not in the option's name. Change this by adding the "Internal"
suffix to the option name in general, rather than just the help text.

Fixes: QTBUG-122774
Pick-to: 6.7
Change-Id: Iaffcde4768a8bf70ba4b8e52cec4ea6490e20bfc
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
(cherry picked from commit 7c218d20aedd8aaf2c0f2b6954ac400902ae040b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>