Bootstrap: remove the UTF-16 and UTF-32 codecs
Unlike most of everything else in the Bootstrap lib, this is code that
couldn't be eliminated by the linker because they were referenced in one
static array. Maybe an exceptionally smart whole-program analysis could
do it, but GCC and Clang LTO modes don't do that now.

I removed the code that performed detection from HTML and from data too.
I could have left the detection of UTF-8 and "other" but this code
wasn't necessary. In particular, QTextStream couldn't benefit from it
because it already defaults to UTF-8, so the detection code would never
determine anything different from the input.

Drive-by removed QStringConverter::availableCodecs() too because it was
in the middle of functions #ifdef'ed out to.

This reduced the size of release-mode moc
   text    data     bss     dec     hex filename
1079858    5440     640 1085938  1091f2 original/moc
1074386    5200     640 1080226  107ba2 updated/moc
  -5472    -240       0   -5712         difference

Change-Id: I01ec3c774d9943adb903fffd17b7f114c42874ac
Reviewed-by: Lars Knoll <lars@knoll.priv.no>